Revert "fix /dev/ptmx mount on chroot"
The entire purpose of the test script debian/tests/fake/schroot-1.6.10-3
is that it behaves like schroot 1.6.10-3, so any difference in behaviour
between schroot 1.6.10-3 and that script is a bug in that script.

Reverting this also makes the test pass in autopkgtest-virt-qemu (with
/dev managed by udev, /dev/ptmx as a device node with permissions 0666,
and /dev/pts/ptmx as a device node with permissions 000), where it
previously failed since commit add2cf76. If add2cf76 was necessary
to make the tests pass in some container environment, then we need to
special-case that other environment instead.

This partially reverts commit add2cf76 "fix /dev/ptmx mount on chroot
and re-enable autopkgtest".

Closes: #983197
Closes: #983311
Signed-off-by: Simon McVittie <smcv@debian.org>