d/tests: Document expectations for newer container managers
schroot 1.6.10-3 with the default profile had a working /dev/ptmx (mostly
by luck) inside an lxc 2 container, but not in an lxc >= 3 container.
This is really a bug in schroot rather than a bug in debootstrap, and
it isn't a debootstrap regression, so document it as an expected failure
(already reported as schroot bug #983423, with a patch) rather than
failing.

Similarly, schroot 1.6.10-3 with the default profile doesn't get a
working /dev/ptmx inside a Docker container.

Signed-off-by: Simon McVittie <smcv@debian.org>