debootstrap: optionally exclude specific dependencies
debootstrap will always install the first alternative of a
dependency. This setting can be used to exclude such a choice, but one
has to be careful that the dependency is satisfied some other way.

This is intended to be used for the "usrmerge | usr-is-merged"
dependency where debootstrap already installs "usr-is-merged" and
should not pull in "usrmerge" as well.

Note that this patch changes the output of "debootstrap --print-debs"
slightly: required packages are separated by space instead of newline
as was already the case for base packages.