functions: Default to merged-/usr, even for newer suites' buildds
For Debian 12 'bookworm' and up, the only supported filesystem layout
is merged /usr. However, as a special exception to facilitate the
transition to that layout, we use the unsupported non-merged-/usr layout
for official buildds, to try to ensure that packages that might be
needed during the upgrade from Debian 11 'bullseye' are not built in a
way that will make them non-functional on non-merged-/usr systems.

For Debian 13 'trixie' and up, because we do not support skipping a
release during upgrades, all packages may assume a merged-/usr
environment, which means we no longer need to make that exception.

We keep the non-merged-/usr exception in place for bookworm, because
upgrading bullseye systems to any future bookworm point release will
be equally-supported, so the conservative thing to do is to continue to
build bookworm stable and security updates in a non-merged-/usr
environment.