d/tests: Don't assert that systemd-tmpfiles works if it isn't installed
A complete test run in autopkgtest-virt-qemu runs debootstrap twice.
The first time has --include=systemd, so we can count on systemd-tmpfiles
being installed, and therefore we can assert that it worked.

The second time (which we only do on a system that was booted with
systemd and is not a container, i.e. a VM or a dedicated physical
machine), we don't --include=systemd and therefore cannot assume that
systemd-tmpfiles is present. If it isn't, then obviously it's not going
to run, so we can't count on the directories it creates being there.

This fixes a regression in !99 and partially fixes the ability to run
debootstrap's autopkgtests under autopkgtest-virt-qemu.

Fixes: !99
Signed-off-by: Simon McVittie <smcv@debian.org>