Let the buildd variant only install essential, build-essential and apt
This ensures that source packages are not accidentally built in an
environment that also has packages of priority:required installed and
thus miss out on build dependencies.

It also makes debootstrap's idea of what packages are needed for
building source packages the same as other parts of the archive (apt,
dose3, dpkg...).

As an exception to that rule, apt is also installed because buildds use
schroot as the sbuild backend which requires apt inside the chroot for
build dependency installation.

List of known source packages missing a B-D on tzdata and mount:

https://bugs.debian.org/cgi-bin/pkgreport.cgi?users=debian-qa@lists.debian.org;tag=missing-build-depends-on-not-build-essential

See also:

https://salsa.debian.org/installer-team/debootstrap/-/merge_requests/106

Closes: #837060