error out if chroot cannot be found in $PATH
Otherwise, if PATH does not contain /sbin and thus no chroot,
debootstrap will fail with this error message:

E: Unable to execute target architecture

This is because arch-test is run with -c which means that arch-test
tries to run chroot but fails and as a result tells debootstrap that the
architecture cannot be executed.

This misleading error message can be prevented by checking for a working
chroot executable in $PATH early on.