Move "unshared" test from autopkgtest to salsaci pipeline
Running this test as part of autopkgtest would be wrong because
autopkgtests are used to gate transitions to testing. But the
debootstrap autopkgtest does not respect the apt pinning settings set up
in the autopkgtest environment and will create a chroot with different
packages than it should. This means that it tests the wrong thing and
should thus not be used for migration gating.